pyproject.toml
[tool.setuptools]
packages = []

[project]
name = "Mystix"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.5.0",
    "aiohttp>=3.10,<4",
    "jishaku"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "isort",
]

[tool.ruff.lint]
select = [
    "C4",
    "F",
    "G",
    "I",
    "PTH",
    "RUF",
    "SIM",
    "TCH",
    "UP",
    "W",
    "PERF",
    "ANN",
]
ignore = [
    "F401",
    "F402",
    "F403",
    "F405",
    "PERF203",
    "RUF001",
    "RUF009",
    "SIM105",
    "UP034",
    "UP038",
    "ANN401",
    "UP031",
    "PTH123",
]

[tool.ruff]
line-length = 120
exclude = ["venv"]

[tool.ruff.lint.isort]
split-on-trailing-comma = true
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
exclude = ["venv"]
useLibraryCodeForTypes = true
typeCheckingMode = "strict"
reportImportCycles = false
reportPrivateUsage = false